evolve: use smaller timesteps early on
The timestep parameter in the config file is now the maximum timestep. The
simulation starts with 1/1000th that value and double until reaching it. This is
useful because large variations are expected at the beginning of the simulation
depending on initial conditions.

Also updated the baseline output as the results slightly changed.